src/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
